+++ src/versionShit.py
@@ -1,4 +1,4 @@
 from modules.backend.semver import SemVer
 
-lastSemVersion = SemVer(4, 0, 0)
-curSemVersion = SemVer(4, 1, 0)
+lastSemVersion = SemVer(4, 1, 0)
+curSemVersion = SemVer(4, 1, 1)